pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "channels>=4.3.2",
    "channels-redis>=4.3.0",
    "daphne>=4.2.1",
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "django-phonenumber-field[phonenumbers]>=8.4.0",
    "django-rest-framework>=0.1.0",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.29.0",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
]
